# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  git_neolidy_top_neo_flowx_internal_application_tool.CreateConnectorRequest:
    properties:
      auth_config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      auth_type:
        type: string
      config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      endpoint:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    required:
    - endpoint
    - name
    - type
    type: object
  git_neolidy_top_neo_flowx_internal_application_tool.CreateToolRequest:
    properties:
      category:
        type: string
      config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      connector_id:
        type: string
      description:
        type: string
      endpoint:
        type: string
      icon:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    required:
    - name
    - type
    type: object
  git_neolidy_top_neo_flowx_internal_application_tool.UpdateConnectorRequest:
    properties:
      auth_config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      auth_type:
        type: string
      config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      endpoint:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  git_neolidy_top_neo_flowx_internal_application_tool.UpdateToolRequest:
    properties:
      category:
        type: string
      config:
        $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_domain_base.JSON'
      connector_id:
        type: string
      description:
        type: string
      endpoint:
        type: string
      icon:
        type: string
      name:
        type: string
      status:
        type: string
      type:
        type: string
    type: object
  git_neolidy_top_neo_flowx_internal_domain_base.JSON:
    additionalProperties: {}
    type: object
  git_neolidy_top_neo_flowx_pkg_response.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: neolidy@outlook.com
    name: jiangfire
    url: https://github.com/jiangfire/flowx
  description: 企业级智能工具治理与流程编排平台 API 文档
  license:
    name: MIT
    url: https://github.com/jiangfire/flowx/blob/main/LICENSE
  termsOfService: https://github.com/jiangfire/flowx
  title: FlowX API
  version: "1.0"
paths:
...
  /connectors:
    get:
      consumes:
      - application/json
      description: 分页查询连接器列表，支持按类型、状态筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 连接器类型
        in: query
        name: type
        type: string
      - description: 连接器状态
        in: query
        name: status
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 连接器列表
      tags:
      - 工具管理
    post:
      consumes:
      - application/json
      description: 创建新的连接器
      parameters:
      - description: 创建连接器请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_tool.CreateConnectorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建连接器
      tags:
      - 工具管理
  /connectors/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除连接器
      parameters:
      - description: 连接器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除连接器
      tags:
      - 工具管理
    get:
      consumes:
      - application/json
      description: 根据ID获取连接器详细信息
      parameters:
      - description: 连接器ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 连接器详情
      tags:
      - 工具管理
    put:
      consumes:
      - application/json
      description: 根据ID更新连接器信息
      parameters:
      - description: 连接器ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新连接器请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_tool.UpdateConnectorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新连接器
      tags:
      - 工具管理
...
  /tools:
    get:
      consumes:
      - application/json
      description: 分页查询工具列表，支持按类型、状态、分类筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 工具类型
        in: query
        name: type
        type: string
      - description: 工具状态
        in: query
        name: status
        type: string
      - description: 工具分类
        in: query
        name: category
        type: string
      - description: 关键词搜索
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 工具列表
      tags:
      - 工具管理
    post:
      consumes:
      - application/json
      description: 创建新的工具
      parameters:
      - description: 创建工具请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_tool.CreateToolRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 创建工具
      tags:
      - 工具管理
  /tools/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除工具
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 删除工具
      tags:
      - 工具管理
    get:
      consumes:
      - application/json
      description: 根据ID获取工具详细信息
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 工具详情
      tags:
      - 工具管理
    put:
      consumes:
      - application/json
      description: 根据ID更新工具信息
      parameters:
      - description: 工具ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新工具请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/git_neolidy_top_neo_flowx_internal_application_tool.UpdateToolRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 更新工具
      tags:
      - 工具管理
  /tools/export:
    post:
      consumes:
      - application/json
      description: 导出工具列表为Excel文件
      parameters:
      - description: '导出列配置 {columns: []string}'
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 导出工具
      tags:
      - 工具管理
  /tools/export/{task_id}:
    get:
      consumes:
      - application/json
      description: 查询工具导出任务的状态
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 查询导出任务状态
      tags:
      - 工具管理
  /tools/import:
    post:
      consumes:
      - multipart/form-data
      description: 通过Excel文件批量导入工具
      parameters:
      - description: Excel文件
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/git_neolidy_top_neo_flowx_pkg_response.APIResponse'
      security:
      - BearerAuth: []
      summary: 导入工具
      tags:
      - 工具管理
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer Token，格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"